requirements.txt
python-telegram-bot==13.7
selenium==3.141.0
beautifulsoup4==4.9.3
python-dotenv==0.15.0